Revert "Fix linter"
