+++ apps/backend/.eslintrc.js
@@ -1,12 +1,25 @@
 module.exports = {
-  extends: ["@expense-tracker/eslint-config/jest"],
+  parser: '@typescript-eslint/parser',
   parserOptions: {
-    project: './tsconfig.json',
+    project: 'tsconfig.json',
     tsconfigRootDir: __dirname,
+    sourceType: 'module',
   },
-  ignorePatterns: [
-    ".eslintrc.js",
-    "dist/",
-    "generated/",
+  plugins: ['@typescript-eslint/eslint-plugin'],
+  extends: [
+    'plugin:@typescript-eslint/recommended',
+    'plugin:prettier/recommended',
   ],
+  root: true,
+  env: {
+    node: true,
+    jest: true,
+  },
+  ignorePatterns: ['.eslintrc.js'],
+  rules: {
+    '@typescript-eslint/interface-name-prefix': 'off',
+    '@typescript-eslint/explicit-function-return-type': 'off',
+    '@typescript-eslint/explicit-module-boundary-types': 'off',
+    '@typescript-eslint/no-explicit-any': 'off',
+  },
 };
