+++ apps/frontend/package.json
@@ -7,10 +7,7 @@
         "dev": "vite",
         "build": "tsc && vite build",
         "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
-        "preview": "vite preview",
-        "test": "vitest",
-        "test:watch": "vitest --watch",
-        "test:ui": "vitest --ui"
+        "preview": "vite preview"
     },
     "dependencies": {
         "react": "^18.2.0",
@@ -27,21 +24,18 @@
         "tailwind-merge": "^2.0.0"
     },
     "devDependencies": {
-        "@expense-tracker/eslint-config": "*",
         "@types/react": "^18.2.37",
         "@types/react-dom": "^18.2.15",
-        "@typescript-eslint/eslint-plugin": "^7.18.0",
-        "@typescript-eslint/parser": "^7.18.0",
+        "@typescript-eslint/eslint-plugin": "^6.10.0",
+        "@typescript-eslint/parser": "^6.10.0",
         "@vitejs/plugin-react": "^4.1.1",
         "autoprefixer": "^10.4.16",
         "eslint": "^8.53.0",
-        "eslint-plugin-react": "^7.33.2",
         "eslint-plugin-react-hooks": "^4.6.0",
         "eslint-plugin-react-refresh": "^0.4.4",
         "postcss": "^8.4.31",
         "tailwindcss": "^3.3.5",
         "typescript": "^5.2.2",
-        "vite": "^5.0.0",
-        "vitest": "^1.0.0"
+        "vite": "^5.0.0"
     }
 }
