+++ apps/frontend/src/pages/Dashboard.tsx
@@ -5,7 +5,6 @@ export const Dashboard: React.FC = () => {
     <div className="p-6">
       <h1 className="text-2xl font-bold mb-6">Dashboard</h1>
       {/* TODO: Add dashboard content */}
-      <h2>I am a dashboard auto-deployed after last fix</h2>
     </div>
   );
 };
