--- apps/frontend/src/providers/AuthContext.ts
@@ -1,7 +1,0 @@
-import { createContext } from "react";
-
-export interface AuthContextType {
-  // TODO: Define auth context interface
-}
-
-export const AuthContext = createContext<AuthContextType | undefined>(undefined);
