+++ apps/frontend/src/providers/AuthProvider.tsx
@@ -1,6 +1,10 @@
-import React from 'react';
-import { AuthContext, AuthContextType } from './AuthContext';
+import React, { createContext, useContext } from 'react';
 
+interface AuthContextType {
+  // TODO: Define auth context interface
+}
+
+const AuthContext = createContext<AuthContextType | undefined>(undefined);
 
 interface AuthProviderProps {
   children: React.ReactNode;
@@ -15,4 +19,12 @@ export const AuthProvider: React.FC<AuthProviderProps> = ({ children }) => {
       {children}
     </AuthContext.Provider>
   );
+};
+
+export const useAuth = () => {
+  const context = useContext(AuthContext);
+  if (context === undefined) {
+    throw new Error('useAuth must be used within an AuthProvider');
+  }
+  return context;
 };
