--- packages/eslint-config/base.js
@@ -1,33 +1,0 @@
-module.exports = {
-  extends: [
-    "eslint:recommended",
-    "plugin:@typescript-eslint/recommended",
-    "prettier"
-  ],
-  parser: "@typescript-eslint/parser",
-  plugins: ["@typescript-eslint"],
-  parserOptions: {
-    sourceType: "module",
-    ecmaVersion: 2020,
-  },
-  env: {
-    node: true,
-    es6: true,
-  },
-  rules: {
-    "@typescript-eslint/no-unused-vars": ["error", { argsIgnorePattern: "^_" }],
-    "@typescript-eslint/explicit-function-return-type": "off",
-    "@typescript-eslint/explicit-module-boundary-types": "off",
-    "@typescript-eslint/no-explicit-any": "warn",
-    "@typescript-eslint/no-empty-function": "warn",
-    "prefer-const": "error",
-    "no-var": "error",
-  },
-  ignorePatterns: [
-    "dist/",
-    "build/",
-    "node_modules/",
-    "*.js",
-    "*.mjs",
-  ],
-};
