--- packages/eslint-config/index.js
@@ -1,3 +1,0 @@
-module.exports = {
-  extends: ["./base.js"],
-};
