--- packages/eslint-config/node.js
@@ -1,12 +1,0 @@
-module.exports = {
-  extends: ["./base.js"],
-  env: {
-    node: true,
-  },
-  rules: {
-    "@typescript-eslint/interface-name-prefix": "off",
-    "@typescript-eslint/explicit-function-return-type": "off",
-    "@typescript-eslint/explicit-module-boundary-types": "off",
-    "@typescript-eslint/no-explicit-any": "warn",
-  },
-};
