--- packages/eslint-config/package.json
@@ -1,24 +1,0 @@
-{
-  "name": "@expense-tracker/eslint-config",
-  "version": "0.0.0",
-  "private": true,
-  "main": "index.js",
-  "files": [
-    "base.js",
-    "react.js",
-    "node.js",
-    "jest.js",
-    "vitest.js"
-  ],
-  "dependencies": {
-    "@typescript-eslint/eslint-plugin": "^7.18.0",
-    "@typescript-eslint/parser": "^7.18.0",
-    "eslint": "^8.53.0",
-    "eslint-config-prettier": "^9.0.0",
-    "eslint-plugin-prettier": "^5.0.0",
-    "eslint-plugin-react": "^7.33.2",
-    "eslint-plugin-react-hooks": "^4.6.0",
-    "eslint-plugin-react-refresh": "^0.4.4",
-    "prettier": "^3.0.0"
-  }
-}
