--- packages/eslint-config/react.js
@@ -1,31 +1,0 @@
-module.exports = {
-  extends: [
-    "./base.js",
-    "plugin:react/recommended",
-    "plugin:react-hooks/recommended",
-  ],
-  plugins: ["react", "react-hooks", "react-refresh"],
-  env: {
-    browser: true,
-  },
-  parserOptions: {
-    ecmaFeatures: {
-      jsx: true,
-    },
-  },
-  settings: {
-    react: {
-      version: "detect",
-    },
-  },
-  rules: {
-    "react/react-in-jsx-scope": "off", // Not needed in React 17+
-    "react/prop-types": "off", // Using TypeScript for prop validation
-    "react-hooks/rules-of-hooks": "error",
-    "react-hooks/exhaustive-deps": "warn",
-    "react-refresh/only-export-components": [
-      "warn",
-      { allowConstantExport: true },
-    ],
-  },
-};
